feat(web): upload multiple files
