+++ packages_rs/nextclade-web/src/components/FilePicker/UploadBoxCompact.tsx
@@ -1,91 +1,10 @@
-import React, { PropsWithChildren, useMemo, useState } from 'react'
-
-import type { TFunction } from 'i18next'
+import React, { PropsWithChildren, useMemo } from 'react'
 import { Button } from 'reactstrap'
-import { sanitizeError } from 'src/helpers/sanitizeError'
-import styled, { DefaultTheme } from 'styled-components'
-import { FileRejection, useDropzone } from 'react-dropzone'
+import styled from 'styled-components'
 import { useTranslation } from 'react-i18next'
 
-import { theme } from 'src/theme'
+import { getUploadZoneTheme, UploadZoneState, useUploadZone } from 'src/components/FilePicker/useUploadZone'
 import { appendDash } from 'src/helpers/appendDash'
-
-export type UpdateErrorsFunction = (prevErrors: string[]) => string[]
-
-export interface MakeOnDropParams {
-  t: TFunction
-
-  onUpload(file: File): void
-
-  setErrors(updateErrors: string[] | UpdateErrorsFunction): void
-}
-
-export function makeOnDrop({ t, onUpload, setErrors }: MakeOnDropParams) {
-  function handleError(error: Error) {
-    if (error instanceof UploadErrorTooManyFiles) {
-      setErrors((prevErrors) => [...prevErrors, t('Only one file is expected')])
-    } else if (error instanceof UploadErrorUnknown) {
-      setErrors((prevErrors) => [...prevErrors, t('Unknown error')])
-    } else {
-      throw error
-    }
-  }
-
-  async function processFiles(acceptedFiles: File[], rejectedFiles: FileRejection[]) {
-    const nFiles = acceptedFiles.length + rejectedFiles.length
-
-    if (nFiles > 1) {
-      throw new UploadErrorTooManyFiles(nFiles)
-    }
-
-    if (acceptedFiles.length !== 1) {
-      throw new UploadErrorTooManyFiles(acceptedFiles.length)
-    }
-
-    const file = acceptedFiles[0]
-    onUpload(file)
-  }
-
-  async function onDrop(acceptedFiles: File[], rejectedFiles: FileRejection[]) {
-    setErrors([])
-    try {
-      await processFiles(acceptedFiles, rejectedFiles)
-    } catch (error: unknown) {
-      handleError(sanitizeError(error))
-    }
-  }
-
-  // eslint-disable-next-line no-void
-  return (acceptedFiles: File[], rejectedFiles: FileRejection[]) => void onDrop(acceptedFiles, rejectedFiles)
-}
-
-export enum UploadZoneState {
-  normal = 'normal',
-  accept = 'accept',
-  reject = 'reject',
-  hover = 'hover',
-}
-
-class UploadErrorTooManyFiles extends Error {
-  public readonly nFiles: number
-
-  constructor(nFiles: number) {
-    super(`when uploading: one file is expected, but got ${nFiles}`)
-    this.nFiles = nFiles
-  }
-}
-
-class UploadErrorUnknown extends Error {
-  constructor() {
-    super(`when uploading: unknown error`)
-  }
-}
-
-export type UploadZoneElems = keyof typeof theme.uploadZone
-
-export function getUploadZoneTheme(props: { state: UploadZoneState; theme: DefaultTheme }, elem: UploadZoneElems) {
-  return props.theme.uploadZone[elem][props.state]
-}
 
 export const UploadZoneWrapper = styled.div`
   display: flex;
@@ -105,8 +24,8 @@ export const UploadZone = styled.div<{ state: UploadZoneState }>`
   display: flex;
   width: 100%;
   height: 100%;
-  min-height: 57px;
   cursor: pointer;
+  min-height: 57px;
   border-radius: 5px;
   border: ${(props) => getUploadZoneTheme(props, 'border')};
   color: ${(props) => getUploadZoneTheme(props, 'color')};
@@ -146,27 +65,19 @@ export const UploadZoneButton = styled(Button)`
 `
 
 export interface UploaderCompactProps {
-  onUpload(file: File): void
+  onUpload(files: File[]): void
 }
 
 export function UploadBoxCompact({ onUpload, children, ...props }: PropsWithChildren<UploaderCompactProps>) {
   const { t } = useTranslation()
-  const [errors, setErrors] = useState<string[]>([])
-
-  const { getRootProps, getInputProps, isDragActive, isDragAccept, isDragReject } = useDropzone({
-    onDrop: makeOnDrop({ t, onUpload, setErrors }),
-    multiple: false,
+  const { state, errors, hasErrors, getRootProps, getInputProps, isDragActive } = useUploadZone({
+    onUpload,
+    multiple: true,
   })
-
-  const hasErrors = errors.length > 0
 
   if (hasErrors) {
     console.warn(`Errors when uploading:\n${errors.map(appendDash).join('\n')}`)
   }
-
-  let state = UploadZoneState.normal
-  if (isDragAccept) state = UploadZoneState.accept
-  else if (isDragReject) state = UploadZoneState.reject
 
   const normal = useMemo(
     () => (
